+++ back/src/test/java/com/example/cpsplatform/PagingUtilsTest.java
@@ -1,10 +1,9 @@
 package com.example.cpsplatform;
 
+import com.example.cpsplatform.utils.PagingUtils;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class PagingUtilsTest {
